services/db/requirements.txt
# services/db/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.7
alembic==1.11.1
SQLAlchemy==2.0.20